# services/iam/api/openapi.yaml
openapi: 3.1.1
info:
  title: Identity and Access API
  version: 1.0.0
  description: HTTP contract for human identity and platform-wide Access Levels.
servers:
  - url: /
security: []
paths:
  /v1/health/live:
    get:
      operationId: getLiveness
      summary: Report whether the IAM process is running
      responses:
        "200":
          description: The process is running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/health/ready:
    get:
      operationId: getReadiness
      summary: Report whether IAM can serve requests
      responses:
        "200":
          description: IAM and PostgreSQL are ready.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "503":
          description: A required dependency is unavailable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/registrations:
    post:
      operationId: registerUser
      summary: Register a Standard Unverified User
      description: >
        Accepts a canonical Email Address and password. Always assigns Standard
        Access Level. Returns the same generic acceptance response whether the
        Email Address is new or already registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
      responses:
        "202":
          description: The registration request was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          description: The registration request is invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "405":
          $ref: "#/components/responses/MethodNotAllowedPost"
  /v1/email-verifications:
    post:
      operationId: verifyEmailAddress
      summary: Prove control of an Email Address
      description: >
        Consumes a single-use verification challenge and activates an Unverified
        User without creating an Authentication Session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailVerificationRequest"
      responses:
        "204":
          description: The Email Address was verified.
        "400":
          description: The verification challenge is invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "405":
          $ref: "#/components/responses/MethodNotAllowedPost"
  /v1/login:
    post:
      operationId: loginUser
      summary: Authenticate an active User
      description: >
        Accepts a canonical Email Address and password. On success, creates an
        independent Authentication Session and sets host-scoped Secure,
        HttpOnly, SameSite=Lax cookies for the opaque Access Token
        (`iam_access_token`) and Refresh Token (`iam_refresh_token`). Token
        material is never returned in the JSON body. Failures for unknown,
        invalid, Unverified, or Suspended Users share one generic response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "204":
          description: Authentication succeeded and cookies were set.
          headers:
            Set-Cookie:
              description: >
                Access and Refresh Token cookies. Each is Secure, HttpOnly,
                SameSite=Lax, Path=/, and host-scoped (no Domain attribute).
              schema:
                type: string
        "400":
          description: The login request body is invalid.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Authentication failed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "405":
          $ref: "#/components/responses/MethodNotAllowedPost"
components:
  schemas:
    Health:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          const: ok
    Accepted:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          const: accepted
    RegistrationRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          maxLength: 128
    EmailVerificationRequest:
      type: object
      additionalProperties: false
      required:
        - challenge
      properties:
        challenge:
          type: string
          minLength: 1
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 128
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
  responses:
    MethodNotAllowed:
      description: The resource does not support the request method.
      headers:
        Allow:
          description: Methods accepted by this resource.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    MethodNotAllowedPost:
      description: The resource does not support the request method.
      headers:
        Allow:
          description: Methods accepted by this resource.
          schema:
            type: string
            const: POST
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"